# packages/protocol/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "monos boundary probe",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/owner/infrastructure": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfrastructureSnapshot"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/github/repositories": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepositoryPage"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/configuration": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationSnapshot"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/configuration/commands": {
      "post": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationResult"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/configuration/commands/{request_id}": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationResult"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/login": {
      "post": {
        "description": "Same-origin password login; sets Secure HttpOnly __Host-monos cookie.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerSessionStatus"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/session": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerSessionStatus"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/logout": {
      "post": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "description": "Revokes current session and clears its cookie; cookie writes require same Origin.",
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerSessionStatus"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/tasks/{task_id}": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOverview"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/tasks/{task_id}/events": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TaskEventCursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEventPage"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/commands": {
      "post": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerCommandResult"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owner/scopes/{scope_id}/commands/{request_id}": {
      "get": {
        "security": [
          {
            "OwnerAuth": []
          },
          {
            "OwnerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerCommandResult"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands": {
      "post": {
        "security": [
          {
            "AttemptAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands/{request_id}": {
      "get": {
        "security": [
          {
            "AttemptAuth": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runner/inventory": {
      "get": {
        "security": [
          {
            "RunnerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Cursor and snapshot_id must be supplied together; restart the entire scan on snapshot conflict.",
            "schema": {
              "$ref": "#/components/schemas/InventoryCursor"
            }
          },
          {
            "name": "snapshot_id",
            "in": "query",
            "required": false,
            "description": "Cursor and snapshot_id must be supplied together; restart the entire scan on snapshot conflict.",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerInventory"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recovery/attempts/{attempt_id}/commands/{request_id}": {
      "get": {
        "security": [
          {
            "RunnerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recovery/operations/{operation_id}": {
      "get": {
        "security": [
          {
            "RunnerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated response; admission is distinct from completed effects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Deterministic protocol error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CommandResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "status"
        ],
        "properties": {
          "schema_version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "admitted",
              "committed",
              "error",
              "unknown"
            ]
          },
          "request_id": {
            "$ref": "#/components/schemas/Id"
          },
          "operation_id": {
            "$ref": "#/components/schemas/Id"
          },
          "result": {
            "$ref": "#/components/schemas/EffectResult"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        }
      },
...
      "InfrastructureSnapshot": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "runners"
        ],
        "properties": {
          "schema_version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "runners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunnerConfiguration"
            },
            "maxItems": 1024
          }
        }
      },
...
      "GitHubRepositoryPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "repositories"
        ],
        "properties": {
          "schema_version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "repositories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GitHubRepository"
            },
            "maxItems": 100
          },
          "next_page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
          }
        }
      }
    },
    "securitySchemes": {
      "OwnerCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-monos",
        "description": "Secure HttpOnly Owner session; state-changing requests require same Origin."
      },
      "OwnerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque Owner session; neither Runner nor Attempt credentials are accepted."
      },
      "AttemptAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Attempt-scoped credential; never a Runner token."
      },
      "RunnerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Service-private enrolled Runner identity."
      }
    }
  }
}